pyproject.toml
[project]
name = "due-diligence-agent"
version = "0.1.0"
description = "Evidence-first technical due diligence agent for GitHub repositories."
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "pathspec>=0.12",
]

[project.scripts]
dda = "dda.cli:app"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"
